refactor(coordinator): 抽出 qa_session 子模块
机械式行为保持拆分：从 coordinator.rs 搬出划词语音问答会话生命周期 —— QA 录音 error
监听、begin/end_qa_session、错误/静默/取消收尾。与 qa.rs（热键/面板状态）互补。
end_qa_session 跨 await 持/放 qa_state 锁、驱动 SSE 流与窗口焦点 marshaling 原样保留。
